[front] feat: each poll can have its own Home page
