+++ frontend/src/pages/home/videos/ExtensionSection.tsx
